% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_in_snow.R
\name{run_in_snow_require}
\alias{run_in_snow_require}
\title{Load packages from within a snow instance test using require}
\usage{
run_in_snow_require(cluster)
}
\description{
Load packages from within a snow instance test using require
}
\examples{
cluster_req <- snow::makeCluster(1)
run_in_snow(cluster_req)
}
