% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_in_snow.R
\name{run_in_snow_requireNamespace}
\alias{run_in_snow_requireNamespace}
\title{Load packages from within a snow instance test using requireNamespace}
\usage{
run_in_snow_requireNamespace(cluster)
}
\description{
Load packages from within a snow instance test using requireNamespace
}
\examples{
cluster_rns <- snow::makeCluster(1)
run_in_snow_requireNamespace(cluster_rns)
}
